New and improved installation protocol
Builds will no longer pile up every time Roblox updates something,
rather, everything is now installed and handled in one directory.
Anytime an update is made, it will only merge the parts that have
changed, rather than the entire thing. The installation has also been
updated to use a multi-threaded installation so that it will run faster.

As a side effect of these changes, instances of Roblox Studio will now
need to be closed in order for updates to be applied, so you should only
use the gametest servers if you're actually trying to test something new
on there.