Rewrote the installer logic to use Roblox's manifest files.
This change reduces the number of files that have to be downloaded and extracted when the mod manager updates thus reducing the time and bandwidth needed to update the client. The tradeoff is a partially slower initial download, but I think it's worth-while. I have also removed future-is-bright from the branch options, for now at least.