Fixed some Flag Editor bugs, improvements made to Explorer Icon Editor!
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
~~~~ Explorer Icon Editor ~~~~
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

- Added a 'Show Modified Icons' toggle, which switches the
icons in the 'Select Icon' view with their modified icons.

- The state of the toggle is saved, and icons that are changed
have their button's background color blink green for a moment.

- The edit icon button can now be pressed regardless if the
override has been enabled or not. Pressing it will just
automatically enable the override icon checkbox.

- The edit icon button now shows the index of the selected
icon. This should be useful to anyone looking to modify
the ReflectionMetadata to use different icons.

- Fixed a bug where the window would get in the way of
Roblox Studio if the user approved the prompt for it
to be updated.

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
~~~~ Flag Editor ~~~~
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

- Fixed a bug where the search filter wouldn't allow numbers
to be typed in.

- Fixed a bug where the wrong flag would be selected in the
overrides table when adding a new flag override.

- Fixed a bug where the editor would incorrectly assume that
the version of studio from Roblox's servers is the same
version that is installed.

- This would lead to it saving the wrong 'last executed version'
if the user declined to update studio, caching the listed flags
incorrectly until the user manually launched studio again.

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~