Refactored Explorer Icon Editor
Roblox stopped embedding the explorer icon spritesheet in the exe, and instead moved it into the content/textures folder.

To reconcile with this change, I have reworked the explorer icon editor into a sprite manager specifically for that texture. This should allow the tool to remain useful for integrating custom icons into whatever future changes Roblox makes to the texture.