Fix another NRE in GetFastVersionGuid
This fixes the same issue that was attempted to be fixed in the last pull request I did, but the regression was reintroduced, so I made a better fix this time around.

The check to set the branch to Roblox if the setting was null was done after GetFastVersionGuid, causing an exception for new users of the mod manager.